fix: Resolve type errors in preprocessing and spatial parser
- Fix int to double conversion in Point constructor
- Rename Rect class to BoundingRect to avoid conflict with Flutter's Rect
- Update all references to use BoundingRect

+++ lib/utils/advanced_preprocessing.dart
@@ -125,7 +125,7 @@ class AdvancedPreprocessing {
 
       if (luminance < 128) {
         // Found edge
-        return Point(x, y);
+        return Point(x.toDouble(), y.toDouble());
       }
 
       x += dirX * 2;
@@ -137,7 +137,7 @@ class AdvancedPreprocessing {
       }
     }
 
-    return Point(startX, startY);
+    return Point(startX.toDouble(), startY.toDouble());
   }
 
   /// Apply perspective transform to correct skew
